FIX: Add View imports to widget classes and fix long to int conversions
Widget classes:
- Added android.view.View import to WidgetListService, TaskWidgetProvider, TaskWidgetProviderMedium
- This resolves View.VISIBLE and View.GONE symbol errors

NotificationManager:
- Cast task.id from long to int for PendingIntent request code
- Cast task.id from long to int for notification ID
- Required because Android APIs expect int, not long

+++ app/src/main/java/com/aisecretary/taskmaster/util/NotificationManager.java
@@ -92,7 +92,7 @@ public class NotificationManager {
 
         PendingIntent pendingIntent = PendingIntent.getActivity(
             context,
-            task.id,
+            (int) task.id,
             intent,
             PendingIntent.FLAG_IMMUTABLE
         );
@@ -126,7 +126,7 @@ public class NotificationManager {
         // Show notification
         android.app.NotificationManager manager =
             (android.app.NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-        manager.notify(NOTIFICATION_ID_TASK_REMINDER_BASE + task.id, builder.build());
+        manager.notify(NOTIFICATION_ID_TASK_REMINDER_BASE + (int) task.id, builder.build());
     }
 
     /**
